Created an easy way to access the leaflet map object. `yourstorymapobject.map` will give you access to the leaflet map. Closes #124 Closes #132
Regarding issue #132, we are not considering creating another build without Leaflet at this time. One of our main goals of the project is to make StoryMap as easy to use as possible, and having many js files with complicated use cases does not fit within that goal. We will be looking at documenting the build process as well as simplifying it in order to serve needs like this one though.